requirements/base.txt
# Core
Django==4.2.11
djangorestframework==3.14.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Auth & Security
djangorestframework-simplejwt==5.3.1
django[argon2]==4.2.11
cryptography==42.0.5
pyotp==2.9.0

# CORS
django-cors-headers==4.3.1

# RBAC / Throttling
django-ratelimit==4.1.0

# Celery
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Stripe
stripe==8.9.0

# Serialization / Filtering
django-filter==24.1
drf-nested-routers==0.93.5

# Utilities
Pillow==10.2.0
qrcode==7.4.2
Faker==23.3.0

# Email
django-templated-email==3.0.1

# Storage
django-storages==1.14.2
boto3==1.34.51

# Encryption
# (cryptography already included above)

# PDF generation
weasyprint==61.2
reportlab==4.1.0

# HTTP
requests==2.31.0

# Validation
phonenumbers==8.13.31

# API Documentation
drf-spectacular==0.27.1
drf-spectacular[sidecar]==0.27.1
